migrate semantic domain tests
