+++ backend/FwLite/MiniLcm.Tests/MiniLcm.Tests.csproj
@@ -20,6 +20,7 @@
         </PackageReference>
         <PackageReference Include="FluentAssertions" Version="6.12.0"/>
         <PackageReference Include="Soenneker.Utils.AutoBogus" Version="2.1.278"/>
+        <PackageReference Include="System.Linq.Async" Version="6.0.1"/>
     </ItemGroup>
 
     <ItemGroup>
